pyproject.toml
[project]
name = "nlp-vertex-ai"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.56.0",
    "gradio>=6.1.0",
    "python-dotenv>=1.2.1",
]
